server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.19.2"
alembic = "^1.7.5"
tool-kit = {git = "git@github.com:kenny-skaggs/tool_kit.git", rev = "main"}
Flask = "^2.0.2"
Flask-RESTful = "^0.3.9"
Flask-Cors = "^3.0.10"
beautifulsoup4 = "^4.10.0"
flask-praetorian = "^1.3.0"
Flask-SocketIO = "^5.1.1"
simple-websocket = "^0.5.0"
PyMySQL = "^1.0.2"
mysqlclient = "^2.1.0"
requests = "^2.27.1"
oauthlib = "^3.2.0"
pyOpenSSL = "^22.0.0"
google-auth = "^2.6.6"
typer = "^0.4.1"
gunicorn = "^20.1.0"
eventlet = "0.30.2"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
